requirements.txt
fastapi
uvicorn[standard]
pillow
tensorflow-cpu
numpy
python-multipart
scikit-learn
streamlit
requests